:sparkles: Option to prepend the series title to the episode title
Adding an option to take the series title and add that to the episode title. For example, the series is "Honest Trailers" and the episode title from Sonarr is "Alien: Covenant". The resulting episode title to be searched would be "Honest Trailers - Alien: Covenant" which matches the title in YouTube